# Implement Relate Function