--- a/geom/alg_relate.go
+++ b/geom/alg_relate.go
@@ -8,39 +8,32 @@ import (
 // the two geometries relate to each other.
 func Relate(a, b Geometry) (IntersectionMatrix, error) {
 	if a.IsEmpty() || b.IsEmpty() {
-		// TODO: Eliminate duplication here. There might be a more elegant way
-		// to do this.
 		var m IntersectionMatrix
 		m = m.with(imExterior, imExterior, imEntry2)
-		if !b.IsEmpty() {
-			switch b.Dimension() {
-			case 0:
-				m = m.with(imExterior, imInterior, imEntry0)
-				m = m.with(imExterior, imBoundary, imEntryF)
-			case 1:
-				m = m.with(imExterior, imInterior, imEntry1)
-				if !b.Boundary().IsEmpty() {
-					m = m.with(imExterior, imBoundary, imEntry0)
-				}
-			case 2:
-				m = m.with(imExterior, imInterior, imEntry2)
-				m = m.with(imExterior, imBoundary, imEntry1)
-			}
+		if a.IsEmpty() && b.IsEmpty() {
+			return m, nil
 		}
-		if !a.IsEmpty() {
-			switch a.Dimension() {
-			case 0:
-				m = m.with(imInterior, imExterior, imEntry0)
-				m = m.with(imBoundary, imExterior, imEntryF)
-			case 1:
-				m = m.with(imInterior, imExterior, imEntry1)
-				if !a.Boundary().IsEmpty() {
-					m = m.with(imBoundary, imExterior, imEntry0)
-				}
-			case 2:
-				m = m.with(imInterior, imExterior, imEntry2)
-				m = m.with(imBoundary, imExterior, imEntry1)
+
+		var flip bool
+		if b.IsEmpty() {
+			b, a = a, b
+			flip = true
+		}
+		switch b.Dimension() {
+		case 0:
+			m = m.with(imExterior, imInterior, imEntry0)
+			m = m.with(imExterior, imBoundary, imEntryF)
+		case 1:
+			m = m.with(imExterior, imInterior, imEntry1)
+			if !b.Boundary().IsEmpty() {
+				m = m.with(imExterior, imBoundary, imEntry0)
 			}
+		case 2:
+			m = m.with(imExterior, imInterior, imEntry2)
+			m = m.with(imExterior, imBoundary, imEntry1)
+		}
+		if flip {
+			m = m.transpose()
 		}
 		return m, nil
 	}