## Description

- Adds the `Relate` function, which returns an `IntersectionMatrix` showing the relationship between two geometries.

## Check List

Have you:

- Added unit tests? Yes.

- Add cmprefimpl tests? (if appropriate?) Yes.

## Related Issue

- https://github.com/peterstace/simplefeatures/issues/334

## Benchmark Results

As expected, no significant change.

<details>

<summary>Click to expand</summary>

```
name                                                      old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
LineEnvelope/0                                              2.03ns ± 5%    2.01ns ±12%    ~     (p=0.336 n=11+15)
LineEnvelope/1                                              1.32ns ± 4%    1.34ns ± 5%    ~     (p=0.473 n=15+15)
LineEnvelope/2                                              1.65ns ± 3%    1.65ns ± 3%    ~     (p=0.670 n=14+14)
LineEnvelope/3                                              1.49ns ± 1%    1.48ns ± 2%    ~     (p=0.345 n=14+14)
MarshalWKB/polygon/n=10                                      222ns ± 4%     221ns ± 5%    ~     (p=0.360 n=15+15)
MarshalWKB/polygon/n=100                                     430ns ± 5%     425ns ± 4%    ~     (p=0.246 n=14+15)
MarshalWKB/polygon/n=1000                                   2.37µs ± 2%    2.37µs ± 5%    ~     (p=0.798 n=15+15)
MarshalWKB/polygon/n=10000                                  20.5µs ± 7%    20.2µs ± 5%    ~     (p=0.325 n=13+14)
UnmarshalWKB/polygon/n=10                                    393ns ± 2%     391ns ± 8%    ~     (p=0.125 n=14+15)
UnmarshalWKB/polygon/n=100                                   599ns ± 3%     608ns ± 8%    ~     (p=0.299 n=13+15)
UnmarshalWKB/polygon/n=1000                                 2.78µs ± 4%    2.79µs ± 5%    ~     (p=0.870 n=15+15)
UnmarshalWKB/polygon/n=10000                                27.6µs ± 5%    27.1µs ± 3%  -1.92%  (p=0.012 n=14+15)
IntersectsLineStringWithLineString/n=10                     1.70µs ± 9%    1.65µs ± 6%    ~     (p=0.058 n=14+15)
IntersectsLineStringWithLineString/n=100                    23.9µs ± 7%    23.4µs ± 4%    ~     (p=0.175 n=15+14)
IntersectsLineStringWithLineString/n=1000                    248µs ± 5%     247µs ± 3%    ~     (p=0.591 n=15+14)
IntersectsLineStringWithLineString/n=10000                  3.32ms ± 3%    3.34ms ± 9%    ~     (p=0.804 n=14+14)
IntersectsMultiPointWithMultiPoint/n=20                     1.80µs ± 2%    1.80µs ± 5%    ~     (p=0.559 n=13+14)
IntersectsMultiPointWithMultiPoint/n=200                    19.2µs ± 6%    19.2µs ± 3%    ~     (p=0.946 n=14+14)
IntersectsMultiPointWithMultiPoint/n=2000                    189µs ± 2%     189µs ± 5%    ~     (p=1.000 n=13+15)
IntersectsMultiPointWithMultiPoint/n=20000                  2.05ms ± 8%    2.05ms ± 6%    ~     (p=0.949 n=14+15)
PolygonSingleRingValidation/n=10                            2.95µs ± 4%    2.93µs ± 4%    ~     (p=0.322 n=15+13)
PolygonSingleRingValidation/n=100                           42.4µs ± 5%    41.7µs ± 2%    ~     (p=0.277 n=15+12)
PolygonSingleRingValidation/n=1000                           508µs ± 3%     508µs ± 3%    ~     (p=0.910 n=14+14)
PolygonSingleRingValidation/n=10000                         6.72ms ± 3%    6.76ms ± 5%    ~     (p=0.561 n=15+14)
PolygonMultipleRingsValidation/n=4                          8.32µs ± 2%    8.29µs ± 2%    ~     (p=0.533 n=14+14)
PolygonMultipleRingsValidation/n=36                         69.5µs ± 2%    70.2µs ± 3%    ~     (p=0.116 n=13+14)
PolygonMultipleRingsValidation/n=400                         922µs ± 5%     919µs ± 4%    ~     (p=0.793 n=14+13)
PolygonMultipleRingsValidation/n=4096                       12.0ms ± 6%    11.9ms ± 7%    ~     (p=0.477 n=15+14)
PolygonZigZagRingsValidation/n=10                           13.7µs ± 3%    13.7µs ± 4%    ~     (p=0.675 n=15+15)
PolygonZigZagRingsValidation/n=100                           160µs ± 5%     159µs ± 4%    ~     (p=0.583 n=13+14)
PolygonZigZagRingsValidation/n=1000                         1.91ms ± 3%    1.90ms ± 3%    ~     (p=0.715 n=15+14)
PolygonZigZagRingsValidation/n=10000                        25.7ms ± 4%    25.5ms ± 5%    ~     (p=0.350 n=14+13)
PolygonAnnulusValidation/n=10                               4.39µs ± 4%    4.43µs ± 7%    ~     (p=0.389 n=15+15)
PolygonAnnulusValidation/n=100                              40.3µs ± 3%    40.5µs ± 6%    ~     (p=0.870 n=15+15)
PolygonAnnulusValidation/n=1000                              699µs ± 3%     701µs ± 3%    ~     (p=0.793 n=14+13)
PolygonAnnulusValidation/n=10000                            8.34ms ± 3%    8.34ms ± 3%    ~     (p=0.892 n=15+13)
MultipolygonValidation/n=1                                   468ns ± 4%     469ns ± 4%    ~     (p=0.812 n=14+14)
MultipolygonValidation/n=4                                  1.08µs ± 3%    1.08µs ± 3%    ~     (p=0.205 n=15+14)
MultipolygonValidation/n=16                                 4.33µs ± 3%    4.31µs ± 2%    ~     (p=0.848 n=14+14)
MultipolygonValidation/n=64                                 19.9µs ± 5%    19.8µs ± 3%    ~     (p=0.527 n=14+13)
MultipolygonValidation/n=256                                 131µs ± 5%     130µs ± 2%    ~     (p=0.949 n=15+14)
MultipolygonValidation/n=1024                                608µs ± 5%     604µs ± 2%    ~     (p=0.541 n=14+14)
MultiPolygonTwoCircles/n=10                                 4.17µs ± 4%    4.26µs ±10%    ~     (p=0.153 n=14+14)
MultiPolygonTwoCircles/n=100                                46.3µs ± 3%    46.4µs ± 5%    ~     (p=0.550 n=13+14)
MultiPolygonTwoCircles/n=1000                                516µs ± 3%     515µs ± 4%    ~     (p=0.734 n=14+14)
MultiPolygonTwoCircles/n=10000                              8.46ms ± 4%    8.68ms ±12%    ~     (p=0.169 n=13+14)
MultiPolygonMultipleTouchingPoints/n=1                      5.83µs ± 5%    5.80µs ± 4%    ~     (p=0.468 n=15+15)
MultiPolygonMultipleTouchingPoints/n=10                     48.5µs ± 8%    47.9µs ± 3%    ~     (p=0.400 n=14+15)
MultiPolygonMultipleTouchingPoints/n=100                     569µs ± 3%     574µs ± 4%    ~     (p=0.174 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1000                   7.16ms ± 5%    7.21ms ± 6%    ~     (p=0.482 n=14+14)
WKTParsing/point                                            2.06µs ± 4%    2.05µs ± 4%    ~     (p=0.517 n=13+15)
DistancePolygonToPolygonOrdering/n=100_swap=false           66.3µs ± 3%    66.4µs ± 2%    ~     (p=0.880 n=14+15)
DistancePolygonToPolygonOrdering/n=100_swap=true            67.3µs ± 5%    67.4µs ± 5%    ~     (p=0.880 n=14+15)
DistancePolygonToPolygonOrdering/n=1000_swap=false          1.06ms ± 2%    1.06ms ± 3%    ~     (p=0.713 n=15+15)
DistancePolygonToPolygonOrdering/n=1000_swap=true           1.07ms ± 4%    1.07ms ± 3%    ~     (p=0.595 n=15+15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false     6.81µs ± 6%    6.78µs ± 5%    ~     (p=0.454 n=14+14)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true      6.74µs ± 4%    6.73µs ± 3%    ~     (p=0.838 n=14+15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false    74.4µs ± 3%    73.7µs ± 5%    ~     (p=0.134 n=14+15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true     74.7µs ± 9%    73.7µs ± 3%    ~     (p=0.305 n=15+15)
MultiLineStringIsSimpleManyLineStrings/n=100                63.3µs ± 5%    63.5µs ± 2%    ~     (p=0.351 n=15+15)
MultiLineStringIsSimpleManyLineStrings/n=1000                698µs ± 2%     697µs ± 2%    ~     (p=0.751 n=13+15)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
IntersectionWithoutValidation/n=10                          63.9µs ± 8%    63.4µs ± 9%    ~     (p=0.451 n=14+15)
IntersectionWithoutValidation/n=100                          123µs ± 6%     123µs ± 9%    ~     (p=0.486 n=15+15)
IntersectionWithoutValidation/n=1000                         550µs ± 5%     537µs ± 2%  -2.24%  (p=0.046 n=15+13)
IntersectionWithoutValidation/n=10000                       4.62ms ± 6%    4.65ms ± 7%    ~     (p=0.653 n=15+15)
NoOp/n=10                                                   5.01µs ± 7%    4.95µs ± 6%    ~     (p=0.644 n=15+14)
NoOp/n=100                                                  16.3µs ± 3%    16.3µs ± 4%    ~     (p=0.603 n=14+14)
NoOp/n=1000                                                  120µs ± 6%     120µs ± 7%    ~     (p=1.000 n=14+13)
NoOp/n=10000                                                1.28ms ±18%    1.16ms ± 8%    ~     (p=0.156 n=15+13)
pkg:github.com/peterstace/simplefeatures/internal/perf goos:linux goarch:amd64
LineStringIsSimpleCircle/n=10                               2.49µs ± 4%    2.47µs ± 3%    ~     (p=0.369 n=14+13)
LineStringIsSimpleCircle/n=100                              44.7µs ± 4%    43.7µs ± 3%  -2.09%  (p=0.004 n=14+15)
LineStringIsSimpleCircle/n=1000                              481µs ± 4%     474µs ± 2%  -1.56%  (p=0.041 n=15+14)
LineStringIsSimpleCircle/n=10000                            6.56ms ± 4%    6.53ms ± 3%    ~     (p=0.367 n=15+15)
LineStringIsSimpleZigZag/10                                 2.28µs ± 3%    2.27µs ± 2%    ~     (p=0.675 n=15+15)
LineStringIsSimpleZigZag/100                                38.0µs ± 4%    37.9µs ± 4%    ~     (p=0.715 n=15+14)
LineStringIsSimpleZigZag/1000                                451µs ± 2%     450µs ± 2%    ~     (p=0.653 n=15+15)
LineStringIsSimpleZigZag/10000                              6.52ms ± 4%    6.50ms ± 4%    ~     (p=0.683 n=15+14)
SetOperation/n=4/Go_Intersection                            55.2µs ± 3%    55.6µs ± 4%    ~     (p=0.287 n=13+13)
SetOperation/n=4/Go_Difference                              56.8µs ± 4%    56.7µs ± 1%    ~     (p=0.742 n=14+12)
SetOperation/n=4/Go_SymmetricDifference                     71.7µs ± 3%    71.8µs ± 3%    ~     (p=0.618 n=15+13)
SetOperation/n=4/Go_Union                                   58.4µs ± 3%    58.8µs ± 4%    ~     (p=0.496 n=13+15)
SetOperation/n=4/GEOS_Intersection                          63.5µs ± 6%    63.5µs ± 7%    ~     (p=1.000 n=14+15)
SetOperation/n=4/GEOS_Difference                            67.2µs ± 8%    66.7µs ± 5%    ~     (p=0.751 n=15+13)
SetOperation/n=4/GEOS_SymmetricDifference                   95.0µs ± 7%    93.8µs ± 5%    ~     (p=0.367 n=15+15)
SetOperation/n=4/GEOS_Union                                 69.3µs ±18%    66.6µs ± 4%    ~     (p=0.172 n=15+14)
SetOperation/n=8/Go_Intersection                            70.3µs ± 3%    70.0µs ± 5%    ~     (p=0.512 n=15+15)
SetOperation/n=8/Go_Difference                              70.3µs ± 4%    70.3µs ± 4%    ~     (p=0.813 n=14+15)
SetOperation/n=8/Go_SymmetricDifference                     90.5µs ± 2%    90.6µs ± 3%    ~     (p=0.867 n=12+15)
SetOperation/n=8/Go_Union                                   74.0µs ± 5%    72.7µs ± 3%    ~     (p=0.050 n=15+15)
SetOperation/n=8/GEOS_Intersection                          75.2µs ± 5%    73.9µs ± 3%    ~     (p=0.246 n=14+14)
SetOperation/n=8/GEOS_Difference                            76.2µs ± 7%    75.3µs ± 4%    ~     (p=0.451 n=14+15)
SetOperation/n=8/GEOS_SymmetricDifference                    108µs ± 6%     110µs ± 6%    ~     (p=0.432 n=12+14)
SetOperation/n=8/GEOS_Union                                 76.5µs ± 8%    76.2µs ± 6%    ~     (p=0.867 n=13+14)
SetOperation/n=16/Go_Intersection                           95.9µs ± 3%    96.9µs ± 4%    ~     (p=0.123 n=15+14)
SetOperation/n=16/Go_Difference                              102µs ± 4%     101µs ± 2%    ~     (p=0.146 n=14+15)
SetOperation/n=16/Go_SymmetricDifference                     137µs ± 3%     137µs ± 3%    ~     (p=0.839 n=14+14)
SetOperation/n=16/Go_Union                                   106µs ± 4%     107µs ± 3%    ~     (p=0.603 n=14+14)
SetOperation/n=16/GEOS_Intersection                         82.4µs ± 5%    81.3µs ± 3%    ~     (p=0.185 n=15+13)
SetOperation/n=16/GEOS_Difference                           90.0µs ± 7%    89.3µs ± 3%    ~     (p=0.624 n=15+15)
SetOperation/n=16/GEOS_SymmetricDifference                   139µs ± 4%     138µs ± 3%    ~     (p=0.300 n=15+12)
SetOperation/n=16/GEOS_Union                                93.2µs ± 8%    92.9µs ± 8%    ~     (p=0.902 n=15+15)
SetOperation/n=32/Go_Intersection                            166µs ± 5%     166µs ± 6%    ~     (p=0.683 n=13+15)
SetOperation/n=32/Go_Difference                              172µs ± 3%     173µs ± 4%    ~     (p=0.653 n=15+15)
SetOperation/n=32/Go_SymmetricDifference                     229µs ± 3%     229µs ± 6%    ~     (p=0.683 n=15+14)
SetOperation/n=32/Go_Union                                   178µs ± 3%     178µs ± 3%    ~     (p=0.751 n=15+13)
SetOperation/n=32/GEOS_Intersection                          100µs ±11%      98µs ± 2%    ~     (p=0.306 n=14+14)
SetOperation/n=32/GEOS_Difference                            111µs ± 7%     109µs ± 4%    ~     (p=0.270 n=15+14)
SetOperation/n=32/GEOS_SymmetricDifference                   182µs ± 4%     182µs ± 3%    ~     (p=0.496 n=14+13)
SetOperation/n=32/GEOS_Union                                 116µs ± 7%     115µs ± 4%    ~     (p=0.856 n=15+13)
SetOperation/n=64/Go_Intersection                            277µs ± 3%     275µs ± 5%    ~     (p=0.201 n=15+14)
SetOperation/n=64/Go_Difference                              295µs ± 4%     291µs ± 3%  -1.35%  (p=0.050 n=14+14)
SetOperation/n=64/Go_SymmetricDifference                     393µs ± 3%     394µs ± 4%    ~     (p=0.838 n=15+15)
SetOperation/n=64/Go_Union                                   308µs ± 4%     309µs ± 5%    ~     (p=0.910 n=14+14)
SetOperation/n=64/GEOS_Intersection                          130µs ± 8%     128µs ± 6%    ~     (p=0.400 n=15+14)
SetOperation/n=64/GEOS_Difference                            153µs ± 3%     155µs ± 4%    ~     (p=0.331 n=14+15)
SetOperation/n=64/GEOS_SymmetricDifference                   274µs ± 5%     275µs ± 3%    ~     (p=0.376 n=14+14)
SetOperation/n=64/GEOS_Union                                 170µs ± 4%     168µs ± 3%    ~     (p=0.294 n=15+13)
SetOperation/n=128/Go_Intersection                           518µs ± 7%     510µs ± 3%    ~     (p=0.187 n=15+15)
SetOperation/n=128/Go_Difference                             535µs ± 2%     536µs ± 3%    ~     (p=0.914 n=15+14)
SetOperation/n=128/Go_SymmetricDifference                    715µs ± 5%     720µs ± 3%    ~     (p=0.250 n=15+15)
SetOperation/n=128/Go_Union                                  554µs ± 2%     558µs ± 8%    ~     (p=0.902 n=15+15)
SetOperation/n=128/GEOS_Intersection                         192µs ± 5%     190µs ± 3%    ~     (p=0.217 n=15+15)
SetOperation/n=128/GEOS_Difference                           228µs ± 8%     225µs ± 4%    ~     (p=0.325 n=15+15)
SetOperation/n=128/GEOS_SymmetricDifference                  432µs ± 6%     426µs ± 3%    ~     (p=0.316 n=15+13)
SetOperation/n=128/GEOS_Union                                246µs ± 3%     243µs ± 2%  -1.27%  (p=0.023 n=15+14)
SetOperation/n=256/Go_Intersection                           935µs ± 3%     931µs ± 5%    ~     (p=0.329 n=14+14)
SetOperation/n=256/Go_Difference                            1.03ms ± 2%    1.03ms ± 4%    ~     (p=0.982 n=14+14)
SetOperation/n=256/Go_SymmetricDifference                   1.39ms ± 2%    1.40ms ± 3%    ~     (p=0.290 n=14+15)
SetOperation/n=256/Go_Union                                 1.07ms ± 3%    1.07ms ± 3%    ~     (p=0.567 n=15+15)
SetOperation/n=256/GEOS_Intersection                         281µs ± 3%     278µs ± 3%    ~     (p=0.104 n=14+14)
SetOperation/n=256/GEOS_Difference                           364µs ± 8%     362µs ± 3%    ~     (p=0.561 n=14+15)
SetOperation/n=256/GEOS_SymmetricDifference                  794µs ± 4%     823µs ± 9%    ~     (p=0.074 n=15+15)
SetOperation/n=256/GEOS_Union                                415µs ± 3%     416µs ± 8%    ~     (p=0.964 n=15+13)
SetOperation/n=512/Go_Intersection                          1.92ms ± 6%    1.97ms ±11%    ~     (p=0.325 n=15+15)
SetOperation/n=512/Go_Difference                            2.03ms ± 6%    2.04ms ± 9%    ~     (p=0.983 n=14+15)
SetOperation/n=512/Go_SymmetricDifference                   2.74ms ± 5%    2.79ms ± 5%    ~     (p=0.146 n=14+15)
SetOperation/n=512/Go_Union                                 2.06ms ± 2%    2.09ms ± 6%    ~     (p=0.158 n=14+15)
SetOperation/n=512/GEOS_Intersection                         498µs ± 5%     504µs ±10%    ~     (p=0.561 n=15+14)
SetOperation/n=512/GEOS_Difference                           658µs ± 5%     663µs ± 7%    ~     (p=0.747 n=15+14)
SetOperation/n=512/GEOS_SymmetricDifference                 1.49ms ± 5%    1.51ms ± 6%    ~     (p=0.126 n=15+15)
SetOperation/n=512/GEOS_Union                                753µs ± 4%     756µs ± 8%    ~     (p=0.981 n=13+14)
SetOperation/n=1024/Go_Intersection                         3.80ms ± 3%    3.82ms ± 3%    ~     (p=0.252 n=15+14)
SetOperation/n=1024/Go_Difference                           3.98ms ± 3%    3.97ms ± 5%    ~     (p=0.325 n=15+15)
SetOperation/n=1024/Go_SymmetricDifference                  5.65ms ± 8%    5.67ms ± 8%    ~     (p=0.838 n=15+15)
SetOperation/n=1024/Go_Union                                4.15ms ± 2%    4.19ms ± 5%    ~     (p=0.734 n=14+14)
SetOperation/n=1024/GEOS_Intersection                        915µs ± 2%     947µs ±14%    ~     (p=0.116 n=13+14)
SetOperation/n=1024/GEOS_Difference                         1.35ms ± 5%    1.36ms ± 9%    ~     (p=0.830 n=13+14)
SetOperation/n=1024/GEOS_SymmetricDifference                3.08ms ± 6%    3.08ms ± 8%    ~     (p=0.603 n=14+14)
SetOperation/n=1024/GEOS_Union                              1.59ms ± 3%    1.58ms ± 4%    ~     (p=0.804 n=14+14)
SetOperation/n=2048/Go_Intersection                         7.80ms ± 5%    7.83ms ± 4%    ~     (p=0.935 n=15+15)
SetOperation/n=2048/Go_Difference                           8.52ms ± 3%    8.52ms ± 2%    ~     (p=0.983 n=15+14)
SetOperation/n=2048/Go_SymmetricDifference                  11.9ms ± 2%    12.0ms ± 5%    ~     (p=0.186 n=13+13)
SetOperation/n=2048/Go_Union                                8.83ms ± 6%    8.87ms ±11%    ~     (p=0.806 n=15+15)
SetOperation/n=2048/GEOS_Intersection                       1.91ms ± 4%    1.92ms ± 7%    ~     (p=0.511 n=14+14)
SetOperation/n=2048/GEOS_Difference                         2.50ms ± 4%    2.52ms ± 5%    ~     (p=0.512 n=15+15)
SetOperation/n=2048/GEOS_SymmetricDifference                5.87ms ± 4%    5.93ms ± 3%    ~     (p=0.112 n=14+15)
SetOperation/n=2048/GEOS_Union                              2.95ms ± 5%    2.95ms ± 4%    ~     (p=0.949 n=14+15)
SetOperation/n=4096/Go_Intersection                         17.0ms ± 6%    17.0ms ± 6%    ~     (p=1.000 n=15+15)
SetOperation/n=4096/Go_Difference                           17.8ms ± 4%    17.9ms ± 5%    ~     (p=0.400 n=15+14)
SetOperation/n=4096/Go_SymmetricDifference                  24.7ms ± 4%    24.9ms ± 5%    ~     (p=0.621 n=14+15)
SetOperation/n=4096/Go_Union                                18.6ms ± 3%    18.5ms ± 5%    ~     (p=0.310 n=15+14)
SetOperation/n=4096/GEOS_Intersection                       3.52ms ± 4%    3.52ms ± 3%    ~     (p=0.780 n=14+15)
SetOperation/n=4096/GEOS_Difference                         5.48ms ± 7%    5.40ms ± 4%    ~     (p=0.331 n=15+14)
SetOperation/n=4096/GEOS_SymmetricDifference                13.7ms ± 7%    13.8ms ± 7%    ~     (p=0.533 n=14+15)
SetOperation/n=4096/GEOS_Union                              6.27ms ± 4%    6.31ms ± 6%    ~     (p=0.591 n=15+14)
SetOperation/n=8192/Go_Intersection                         35.2ms ± 9%    35.2ms ± 6%    ~     (p=0.744 n=15+15)
SetOperation/n=8192/Go_Difference                           36.7ms ± 6%    37.0ms ±10%    ~     (p=0.847 n=15+14)
…